+++ app.json
@@ -25,6 +25,6 @@
     },
     "web": {
       "favicon": "./assets/favicon.png"
-    }
-  }
+    } 
+   }
 }
