+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "swarm-ios-app",
   "version": "1.0.0",
-  "main": "node_modules/expo/AppEntry.js",
+  "main": "src/App.tsx",
   "scripts": {
     "start": "expo start",
     "android": "expo start --android",
@@ -10,11 +10,14 @@
   },
   "dependencies": {
     "@expo/metro-runtime": "~3.1.3",
+    "@react-navigation/native": "^6.1.17",
     "expo": "~50.0.14",
     "expo-status-bar": "~1.11.1",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-native": "0.73.6",
+    "react-native-safe-area-context": "^4.8.2",
+    "react-native-screens": "~3.29.0",
     "react-native-web": "~0.19.6"
   },
   "devDependencies": {
