+++ src/App.tsx
@@ -1,0 +1,27 @@
+import { registerRootComponent } from 'expo';
+import { StatusBar } from 'expo-status-bar';
+import React from 'react';
+import { NavigationContainer } from '@react-navigation/native';
+import { StyleSheet, Text, View } from 'react-native';
+
+const App = () => {
+  return (
+    <NavigationContainer>
+      <View style={styles.container}>
+        <Text>Swarm Beekeeping App</Text>
+        <StatusBar style="auto" />
+      </View>
+    </NavigationContainer >
+  );
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#fff',
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+});
+
+export default registerRootComponent(App);
